mongodoc.CheckCharmResource -> charmstore.checkCharmResource.
